Add files via upload
Learnt something new in Basic JS today, Learnt some boxes, which we usually see in the websites, when we click on some link. like they ask us whether we want to leabe the website ? or somwthing like before entering to this site tell us the answer of this question or it tells us that we are leaving the website. And with these boxes, I've learnt something aboyt switch cases in which we can add multiple cases and either one of case can be true and if none of the case is true than we can put the default case.